:truck: Move `docs_src` helper methods to sole dependent module
